% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-functions.R
\name{abc}
\alias{abc}
\title{Main function for fitting model by ABC}
\usage{
abc(observed_sequence_file, simulated_sequence_file, simulated_param_file,
  match_tol = 0.01)
}
\arguments{
\item{observed_sequence_file}{A string, describing a fasta file
that contains the observed sequences.}
}
\value{
A matrix, columns representing parameters, of accepted parameters.
}
\description{
Main function for fitting model by ABC
}
