% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-functions.R
\name{abc_rej}
\alias{abc_rej}
\title{Main function for fitting model by ABC}
\usage{
abc_rej(sim_params, sim_statistics, obs_statistics, match_tol = 0.01)
}
\arguments{
\item{sim_params}{Parameters used in the simulator.}

\item{sim_statistics}{Statistics corresponding to each parameter set.}

\item{obs_statistics}{Statistics for the observed data.}

\item{match_tol}{What fraction of the samples to keep.}
}
\value{
A matrix, columns representing parameters, of accepted parameters.
}
\description{
Main function for fitting model by ABC
}
