% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-functions.R
\name{check_abc_on_sims}
\alias{check_abc_on_sims}
\title{Performs ABC on a simulated data set}
\usage{
check_abc_on_sims(params, statistics, match_tol)
}
\arguments{
\item{params}{A matrix, nsims x nparams, giving the parameter
values used in the simulations.}

\item{statistics}{A matrix, nsims x nstats, giving the summary
statistics computed from each of the simulations.}

\item{match_tol}{Keep the closest match_tol fraction of the
parameters for the posterior.}
}
\value{
A data frame with columns for the true value of the
parameter, the posterior mean of the parameter, and the name of the
parameter.
}
\description{
Performs ABC on a simulated data set
}
