% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-functions.R
\name{draw_simulated_sequences}
\alias{draw_simulated_sequences}
\title{Interface to the C++ simulator}
\usage{
draw_simulated_sequences(n_seqs)
}
\arguments{
\item{n_seqs}{The number of sequences to draw.}
}
\value{
A character vector, length equal to n_seqs, each element
equal to one draw from the simulator.
}
\description{
Interface to the C++ simulator
}
