% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{generate_mmr_samples}
\alias{generate_mmr_samples}
\title{Sample MMR tract sizes}
\usage{
generate_mmr_samples(n, p, b)
}
\arguments{
\item{n}{The number of samples.}

\item{p}{The parameter for the geometric distribution.}

\item{b}{The distance between the AID lesions.}
}
\value{
A data frame with columns corresponding to a, b, c
    (distance to the left of the lesions, distance between lesions,
    distance to the right of th elesions) and rows corresponding to
    samples.
}
\description{
Given a distribution for EXO stripping on either side and a
distance between AID lesions, use rejection sampling to sample from
the conditional distribution given that the amount of exo stripping
resulted in a single tract.
}
