% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{get_gr}
\alias{get_gr}
\title{Something something}
\usage{
get_gr(germline_seq, mutated_seqs, r_max, r_min, mutation_probs = NULL)
}
\arguments{
\item{germline_seq}{The germline sequence.}

\item{mutated_seqs}{The mutated sequences.}

\item{r_max}{Maximum distance between mutations to calculate g(r)
for.}

\item{r_min}{Minimum distance between mutations to calculate g(r)
for.}

\item{mutation_freqs}{If NULL, will estiamte the per-position
mutation probabilities from germline_seq and
mutated_seqs. Otherwise, mutation_freqs is a vector where
mutation_freqs[i] gives the probability of mutation at position
i.}
}
\value{
A data frame with columns dist (the distance between
    mutations) and gr (spatial colocalization statistic for the
    corresponding distance between mutations.)
}
\description{
Something something
}
