% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{get_mutation_freqs}
\alias{get_mutation_freqs}
\title{Finds per-position mutation frequences for a set of sequences}
\usage{
get_mutation_freqs(germline_seq, mutated_seqs)
}
\arguments{
\item{germline_seq}{Germline sequence as a character vector.}

\item{mutated_seqs}{A list of mutated sequences, each one as a
character vector.}
}
\value{
A data frame with one column for position, one column for
    mutation count, one column for mutation frequency, and one
    column for germline base.
}
\description{
Finds per-position mutation frequences for a set of sequences
}
