% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{get_mutation_indices}
\alias{get_mutation_indices}
\title{Finds the positions of mutations}
\usage{
get_mutation_indices(seq1, seq2, as_indicator = TRUE)
}
\arguments{
\item{germline_seq}{Germline sequence as a character vector.}

\item{mutated_seq}{Mutated sequence as a character vector.}
}
\value{
If as_vector, an indicator vector of mutation at each
    position. Otherwise a vector containing the indices of the
    mutations. The indices of the mutations.
}
\description{
Finds the positions of mutations
}
