% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{get_num_mutations}
\alias{get_num_mutations}
\title{Finds the number of mutations.}
\usage{
get_num_mutations(mutated_seq, germline_seq)
}
\arguments{
\item{mutated_seq}{String giving the mutated sequence.}

\item{germline_seq}{String giving the germline sequence germline
sequence.}
}
\value{
The number of mutations.
}
\description{
Finds the number of mutations.
}
