% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{get_pair_frequency}
\alias{get_pair_frequency}
\title{Finds the frequencies of all pairs of mutations within a certain distance.}
\usage{
get_pair_frequency(mutations, npositions, max_dist, min_dist = 0,
  mutation_probs = NULL)
}
\arguments{
\item{mutations}{A list, each element of which is a vector
containing the positions of mutations in the corresponding
sequence.}

\item{npositions}{The number of positions (length of the germline
sequence).}

\item{max_dist}{The maximum distance to compute.}

\item{min_dist}{The minimum distance to compute.}

\item{mutation_probs}{mutation_probs[i] gives the probability of
mutation at position i.}
}
\description{
The idea here is to compare the number of times we see each pair of
mutations with the number of times we would expect to see it under
an independence model. Pairs of mutations are described as a pair
(i,j), with i < j. If mutations are independent, the probability of
seeing mutations i and j in a sequence with n mutations total is n
* (n-1) * p(i) * p(j), where p(i) and p(j) are the marginal
probabilities of mutations at positions i and j. Therefore, the
expected number of times we see mutation pair (i,j) in the set of
sequences is p(i) * p(j) * sum_seq n_seq (n_seq - 1).
}
