% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{make_pol_eta_string}
\alias{make_pol_eta_string}
\title{String representation of pol eta params}
\usage{
make_pol_eta_string(substitution_matrix, nucleotides)
}
\arguments{
\item{substitution_matrix}{A 4x4 matrix, rows as germline bases,
columns as target bases, with substitution probabilities.}

\item{nucleotides}{A vector giving the nucleotides.}
}
\value{
A string that can be turned into a python dictionary with
    ast.literal_eval.
}
\description{
Creates a string that, when loaded into python with
ast.literal_eval, will make a dictionary with nucleotides as keys
and substitution probabilities as values.
}
