% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{make_predictors_from_seqs}
\alias{make_predictors_from_seqs}
\title{Make a predictor matrix from a set of sequences}
\usage{
make_predictors_from_seqs(seqs)
}
\arguments{
\item{seqs}{A vector, length equal to the number of sequences, each
element a string giving the sequence.}
}
\value{
A matrix, number of rows equal to length(seqs), each column
a predictor.
}
\description{
Presumably to be used as input to a neural net
}
