% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-functions.R
\name{make_summary_stats}
\alias{make_summary_stats}
\title{Creates summary statistics from a set of sequences.}
\usage{
make_summary_stats(seqs)
}
\arguments{
\item{seqs}{A character vector, each element a sequence, to compute
summary statistics on.}
}
\value{
A vector containing summary statistics.
}
\description{
Creates summary statistics from a set of sequences.
}
