% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-functions.R
\name{match_statistics}
\alias{match_statistics}
\title{Finds the simulated statistics that match most closely to the
observed statistics.}
\usage{
match_statistics(observed_stats, simulation_stats, tol)
}
\arguments{
\item{observed_stats}{The observed statistics, a vector of length
n_stats.}

\item{simulation_stats}{Statistics from the simulator, a matrix of
size n_sims by n_stats.}

\item{tol}{The fraction of the draws to retain. Total will be
n_sims * tol.}
}
\description{
Finds the simulated statistics that match most closely to the
observed statistics.
}
