% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{mutation_rates}
\alias{mutation_rates}
\title{Computes per-base mutation rates for each germline sequence in a sequence set}
\usage{
mutation_rates(seq_set)
}
\arguments{
\item{seq_set}{seq_set object.}
}
\value{
A data frame: One column gives the germline sequence id,
    one column gives the position within the germline sequence, one
    column gives the mutation rate at that position in the germline
    sequence, and one column gives the germline base.
}
\description{
Computes per-base mutation rates for each germline sequence in a sequence set
}
