% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor-creation-functions.R
\name{one_hot_1d_sequences}
\alias{one_hot_1d_sequences}
\title{Creates a matrix of one-hot representations of sequences.}
\usage{
one_hot_1d_sequences(seqs)
}
\arguments{
\item{seqs}{A vector of strings, each one a sequence.}
}
\value{
A matrix, number of rows equal to the number of sequences,
number of columns four times the length of each sequence. Each row
is a vectorized version of the output from sequence_to_matrix.
}
\description{
Creates a matrix of one-hot representations of sequences.
}
