% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor-creation-functions.R
\name{one_hot_2d_sequences}
\alias{one_hot_2d_sequences}
\title{Creates an array of one-hot matrix representations of a set of
sequences.}
\usage{
one_hot_2d_sequences(seqs)
}
\arguments{
\item{seqs}{A vector of strings, each one a sequence.}
}
\value{
An array, first dimension equal to the number of sequences,
second dimension equal to the length of each sequence, third
dimension equal to 4 (number of nucleotides). Element (i, j, k) = 1
indicates that the base at the jth position in the ith sequence is
the kth element of the vector (A, T, G, C).
}
\description{
Creates an array of one-hot matrix representations of a set of
sequences.
}
