% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor-creation-functions.R
\name{one_hot_2d_sequences_for_conv_net}
\alias{one_hot_2d_sequences_for_conv_net}
\title{Creates an array of one-hot representations of sequences for use as
input to a convolutional net.}
\usage{
one_hot_2d_sequences_for_conv_net(seqs)
}
\arguments{
\item{seqs}{A vector of strings, each one a sequence.}
}
\value{
The same as one_hot_2d_sequences, but with one extra
dimension.
}
\description{
Creates an array of one-hot representations of sequences for use as
input to a convolutional net.
}
