% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-functions.R
\name{process_fasta}
\alias{process_fasta}
\title{Fasta to character vector.}
\usage{
process_fasta(fasta_file)
}
\arguments{
\item{fasta_file}{The file}
}
\value{
A character vector, each element of the vector one sequence
from the fasta file.
}
\description{
Fasta to character vector.
}
