% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{python_vec_string}
\alias{python_vec_string}
\title{String representation of a vector.}
\usage{
python_vec_string(vec)
}
\arguments{
\item{vec}{The vector to be encoded.}
}
\value{
A string.
}
\description{
Creates a string that, when loaded into python with
ast.literal_eval, will make a list.
}
