% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{seq_set_from_fasta}
\alias{seq_set_from_fasta}
\title{Make SeqSet from fasta file plus germline}
\usage{
seq_set_from_fasta(mutated_seq_fasta, germline_fasta = NULL,
  germline_string = NULL)
}
\arguments{
\item{mutated_seq_fasta}{A fasta file containing the mutated sequences.}

\item{germline_fasta}{A fasta file containing the germline sequence.}

\item{germline_string}{A string giving the germline sequence.}
}
\value{
SeqSet object, contains the mutated sequences and their
    corresponding germlien sequences. Sequences are stored as
    character vectors.
}
\description{
Make SeqSet from fasta file plus germline
}
