% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{seq_set_from_partis}
\alias{seq_set_from_partis}
\title{Make SeqSet from partis output}
\usage{
seq_set_from_partis(partis_df, trim_to_v = FALSE, v_gene_fasta = NULL)
}
\arguments{
\item{partis_file}{}
}
\value{
SeqSet object, contains the mutated sequences and their
    corresponding germline sequences. Sequences are stored as
    character vectors.
}
\description{
Make SeqSet from partis output
}
