% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{sequence_matrix_representation}
\alias{sequence_matrix_representation}
\title{Make a predictor matrix from a sequence}
\usage{
sequence_matrix_representation(seq)
}
\arguments{
\item{seq}{A string, length equal to the number of sequences, each
element a string giving the sequence.}
}
\value{
A matrix, number of rows equal to length(seqs), each column
a predictor.
}
\description{
Presumably to be used as input to a neural net
}
