% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor_creation_functions.R
\name{sequence_to_matrix}
\alias{sequence_to_matrix}
\title{Creates a matrix representation of a sequence}
\usage{
sequence_to_matrix(seq)
}
\arguments{
\item{seq}{A string with the sequence.}
}
\value{
A matrix, number of rows equal to the length of the
sequence, 4 columns. Element (i, j)=1 indicates that the base at
the ith position in the sequence is the jth element of the vector
(A, T, G, C)
}
\description{
Creates a matrix representation of a sequence
}
