% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{shm_sim}
\alias{shm_sim}
\title{Wrapper function for the python simulator}
\usage{
shm_sim(write_summary_stats = TRUE, summary_stat_output = "output",
  write_seqs = FALSE, sequence_output = "output.fasta",
  aid_context_model = "data/logistic_fw_3mer.csv", context_model_length = 3,
  context_model_pos_mutating = 2,
  germline_sequence = "/Users/juliefukuyama/GitHub/HyperMutationModels/data/gpt.fasta",
  ber_lambda = 0.5, bubble_size = 20, exo_left = 20, exo_right = 20,
  n_mutation_rounds = 3, n_seqs = 500, ber_params = c(0.25, 0.25, 0.25,
  0.25), pol_eta_params = diag(rep(1, 4)), p_fw = 0.5)
}
\arguments{
\item{ber_params}{A vector giving substitution rates, nucleotide
order is AGCT.}

\item{pol_eta_params}{A matrix with giving nucleotide substitution
rates, nucleotide order is AGCT, row indicates germline base,
column indicates target base (so the substitution rates
starting from A are given by the first row of the matrix).}
}
\description{
Wrapper function for the python simulator
}
