% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{shm_sim_abc}
\alias{shm_sim_abc}
\title{Wrapper for the python simulator compatible with EasyABC}
\usage{
shm_sim_abc(x)
}
\arguments{
\item{x}{A vector with the parameters.}
}
\value{
A vector of summary statistics.
}
\description{
Wrapper for the python simulator compatible with EasyABC
}
