% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{spatial_colocalization}
\alias{spatial_colocalization}
\title{Computes spatial colocalization statistic}
\usage{
spatial_colocalization(seq_set, r_max = 25, r_min = 0)
}
\arguments{
\item{r_max}{The maximum value of r at which to compute spatial colocalization.}

\item{r_min}{The minimum value of r at which to compute spatial colocalization.}

\item{gl_seq}{A string containing the germline sequence.}

\item{mutated_seqs}{A list of strings containing the mutated sequences.}
}
\description{
Computes spatial colocalization statistic
}
