% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-functions.R
\name{get_net_summary_stats}
\alias{get_net_summary_stats}
\title{Creates summary statistics}
\usage{
get_net_summary_stats(seqs, predictor_creation_fn, net, groups,
  extra_preds = NULL)
}
\arguments{
\item{seqs}{A character vector, length equal to the number of sequences.}

\item{predictor_creation_fn}{A function that takes a sequence
vector and returns a predictor matrix.}

\item{net}{A trained net.}

\item{groups}{A factor describing how the sequences should be
aggregated. If groups of sequences were drawn from the same
parameters, the summary statistics are averaged over each group.}

\item{extra_preds}{Any non-sequence predictors to use.}
}
\value{
A matrix of summary statistics, number of rows equal to the
number of groups, number of columns equal to the number of summary
statistics (the number of parameters the net estimates).
}
\description{
Given a set of sequences, a function that takes a set of sequences
and returns a predictor matrix, and a net, computes the net's
predictions and returns the average of the predictions over the
entire set of sequences.
}
