% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmr-functions.R
\name{pca_filter}
\alias{pca_filter}
\title{PCA filtering of simulations}
\usage{
pca_filter(abc_output, sum_stats, nf, tol, cols = 1:length(sum_stats),
  scale = NULL)
}
\arguments{
\item{abc_output}{The output from ABC_rejection.}

\item{sum_stats}{The observed summary statistic vector.}

\item{nf}{The number of principal components to use for distance computations.}

\item{tol}{The fraction of simulations to keep.}

\item{cols}{The columns of summary statistics on which to perform PCA. The other columns are left unchanged.}
}
\value{
Filtered version
}
\description{
PCA filtering of simulations
}
